Lights are now stored in the model, separated model and view matrices, speed improvements for GL and GLES renderers, added project for remove linux build, updated FMOD

+++ WargameEngine/WargameEngine/WargameEngine/controller/GameController.cpp
@@ -1,6 +1,7 @@
 #include "GameController.h"
 #define _USE_MATH_DEFINES
 #include <math.h>
+#include <float.h>
 #include "ScriptRegisterFunctions.h"
 #include "../model/ObjectGroup.h"
 #include "../model/Object.h"
@@ -143,7 +144,7 @@ bool CGameController::OnRightMouseUp(CVector3f const& begin, CVector3f const& en
 	bool result = false;
 	if (m_rotationPosBegin && object)
 	{
-		float rotation = 90.0f + (atan2(point.y - m_rotationPosBegin->y, point.x - m_rotationPosBegin->x) * 180.0f / (float)M_PI);
+		float rotation = 90.0f + static_cast<float>(atan2(point.y - m_rotationPosBegin->y, point.x - m_rotationPosBegin->x) * 180.0f / (float)M_PI);
 		if (sqrt((point.x - m_rotationPosBegin->x) * (point.x - m_rotationPosBegin->x) + (point.y - m_rotationPosBegin->y) * (point.y - m_rotationPosBegin->y)) > 0.2)
 		{
 			object->Rotate(rotation - rot);
@@ -168,7 +169,7 @@ bool CGameController::OnMouseMove(CVector3f const& begin, CVector3f const& end, 
 	{
 		float rot = selected->GetRotation();
 		auto point = RayToPoint(begin, end);
-		float rotation = 90 + (atan2(point.y - m_rotationPosBegin->y, point.x - m_rotationPosBegin->x) * 180 / (float)M_PI);
+		float rotation = 90 + static_cast<float>(atan2(point.y - m_rotationPosBegin->y, point.x - m_rotationPosBegin->x) * 180.0f / (float)M_PI);
 		if (sqrt((point.x - m_rotationPosBegin->x) * (point.x - m_rotationPosBegin->x) + (point.y - m_rotationPosBegin->y) * (point.y - m_rotationPosBegin->y)) > 0.2)
 			m_model.GetSelectedObject()->Rotate(rotation - rot);
 	}
@@ -582,7 +583,7 @@ void CObjectDecorator::Update(long long timeSinceLastUpdate)
 	}
 	CVector3f dir = m_goTarget - m_object->GetCoords();
 	dir.Normalize();
-	m_object->SetRotation(atan2(dir.y, dir.x) * 180.0f / (float)M_PI);
+	m_object->SetRotation(static_cast<float>(atan2(dir.y, dir.x) * 180.0f / (float)M_PI));
 	dir = dir * static_cast<float>(timeSinceLastUpdate) / 1000.0f * m_goSpeed;
 	if (dir.GetLength() > (m_goTarget - m_object->GetCoords()).GetLength()) dir = (m_goTarget - m_object->GetCoords());
 	m_object->Move(dir.x, dir.y, dir.z);
